Factor out a WalletListFooter component
